数据类型修改.

+++ jkeep-jpa/src/main/java/com/ml/jkeep/jpa/system/entity/sys/Branch.java
@@ -91,6 +91,6 @@ public class Branch {
      * 删除标记
      */
     @Column(name = "d_flag")
-    private Integer dFlag;
+    private Byte dFlag;
 
 }
