+++ jkeep-jpa/src/main/java/com/ml/jkeep/jpa/system/entity/sys/Element.java
@@ -129,6 +129,6 @@ public class Element {
      * 删除标记
      */
     @Column(name = "d_flag")
-    private Integer dFlag;
+    private Byte dFlag;
 
 }
